@@ -11,7 +11,7 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "ajv": "^6.12.4",
+    "ajv": "^6.12.6",
     "ajv-errors": "^1.0.1",
     "ajv-i18n": "^3.5.0",
     "core-js": "^3.6.5",
@@ -41,6 +41,7 @@
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^5.0.2",
     "@vue/test-utils": "2.0.0-beta.5",
+    "babel-eslint": "^10.1.0",
     "circular-dependency-plugin": "^5.2.0",
     "eslint": "^6.7.2",
     "eslint-plugin-prettier": "^3.1.3",
@@ -50,7 +51,7 @@
     "monaco-editor-webpack-plugin": "^1.9.1",
     "prettier": "^1.19.1",
     "rimraf": "^3.0.2",
-    "typescript": "~3.9.3",
+    "typescript": "^4.3.5",
     "vue-jest": "^5.0.0-alpha.4"
   },
   "gitHooks": {
