@@ -1,37 +1,37 @@
-import Ajv from 'ajv'
-import toPath from 'lodash.topath'
-const i18n = require('ajv-i18n') // eslint-disable-line
+import Ajv from "ajv";
+import toPath from "lodash.topath";
+const i18n = require("ajv-i18n"); // eslint-disable-line
 
-import { Schema } from './types'
-import { isObject } from './utils'
+import { Schema } from "./types";
+import { isObject } from "./utils";
 
 interface TransformedErrorObject {
-  name: string
-  property: string
-  message: string | undefined
-  params: Ajv.ErrorParameters
-  schemaPath: string
+  name: string;
+  property: string;
+  message: string | undefined;
+  params: Ajv.ErrorParameters;
+  schemaPath: string;
 }
 
 interface ErrorSchemaObject {
-  [level: string]: ErrorSchema
+  [level: string]: ErrorSchema;
 }
 
 export type ErrorSchema = ErrorSchemaObject & {
-  __errors?: string[]
-}
+  __errors?: string[];
+};
 function toErrorSchema(errors: TransformedErrorObject[]) {
-  if (errors.length < 1) return {}
+  if (errors.length < 1) return {};
 
   return errors.reduce((errorSchema, error) => {
-    const { property, message } = error
-    const path = toPath(property) // /obj/a -> [obj, a]
-    let parent = errorSchema
+    const { property, message } = error;
+    const path = toPath(property); // /obj/a -> [obj, a]
+    let parent = errorSchema;
 
     // If the property is at the root (.level1) then toPath creates
     // an empty array element at the first index. Remove it.
-    if (path.length > 0 && path[0] === '') {
-      path.splice(0, 1)
+    if (path.length > 0 && path[0] === "") {
+      path.splice(0, 1);
     }
 
     // {
@@ -41,29 +41,29 @@ function toErrorSchema(errors: TransformedErrorObject[]) {
     // } // /obj/a
     for (const segment of path.slice(0)) {
       if (!(segment in parent)) {
-        ;(parent as any)[segment] = {}
+        (parent as any)[segment] = {};
       }
-      parent = parent[segment]
+      parent = parent[segment];
     }
 
     if (Array.isArray(parent.__errors)) {
       // We store the list of errors for this node in a property named __errors
       // to avoid name collision with a possible sub schema field named
       // "errors" (see `validate.createErrorHandler`).
-      parent.__errors = parent.__errors.concat(message || '')
+      parent.__errors = parent.__errors.concat(message || "");
     } else {
       if (message) {
-        parent.__errors = [message]
+        parent.__errors = [message];
       }
     }
-    return errorSchema
-  }, {} as ErrorSchema)
+    return errorSchema;
+  }, {} as ErrorSchema);
 }
 
 function transformErrors(
-  errors: Ajv.ErrorObject[] | null | undefined,
+  errors: Ajv.ErrorObject[] | null | undefined
 ): TransformedErrorObject[] {
-  if (errors === null || errors === undefined) return []
+  if (errors === null || errors === undefined) return [];
 
   return errors.map(({ message, dataPath, keyword, params, schemaPath }) => {
     return {
@@ -72,44 +72,44 @@ function transformErrors(
       message,
       params,
       schemaPath,
-    }
-  })
+    };
+  });
 }
 
 export function validateFormData(
   validator: Ajv.Ajv,
   formData: any,
   schema: Schema,
-  locale = 'zh',
-  customValidate?: (data: any, errors: any) => void,
+  locale = "zh",
+  customValidate?: (data: any, errors: any) => void
 ) {
-  let validationError = null
+  let validationError = null;
   try {
-    validator.validate(schema, formData)
+    validator.validate(schema, formData);
   } catch (err) {
-    validationError = err
+    validationError = err;
   }
 
-  i18n[locale](validator.errors)
-  let errors = transformErrors(validator.errors)
+  i18n[locale](validator.errors);
+  let errors = transformErrors(validator.errors);
 
   if (validationError) {
     errors = [
       ...errors,
-      {
-        message: validationError.message,
-      } as TransformedErrorObject,
-    ]
+      // {
+      //   message: validationError.message,
+      // } as TransformedErrorObject,
+    ];
   }
 
-  const errorSchema = toErrorSchema(errors)
+  const errorSchema = toErrorSchema(errors);
 
   if (!customValidate) {
     return {
       errors,
       errorSchema,
       valid: errors.length === 0,
-    }
+    };
   }
 
   /**
@@ -122,56 +122,56 @@ export function validateFormData(
    *
    * raw.obj.a
    */
-  const proxy = createErrorProxy()
-  customValidate(formData, proxy)
-  const newErrorSchema = mergeObjects(errorSchema, proxy, true)
+  const proxy = createErrorProxy();
+  customValidate(formData, proxy);
+  const newErrorSchema = mergeObjects(errorSchema, proxy, true);
 
   return {
     errors,
     errorSchema: newErrorSchema,
     valid: errors.length === 0,
-  }
+  };
 }
 
 function createErrorProxy() {
-  const raw = {}
+  const raw = {};
   return new Proxy(raw, {
     get(target, key, reciver) {
-      if (key === 'addError') {
+      if (key === "addError") {
         return (msg: string) => {
-          const __errors = Reflect.get(target, '__errors', reciver)
+          const __errors = Reflect.get(target, "__errors", reciver);
           if (__errors && Array.isArray(__errors)) {
-            __errors.push(msg)
+            __errors.push(msg);
           } else {
-            ;(target as any).__errors = [msg]
+            (target as any).__errors = [msg];
           }
-        }
+        };
       }
-      const res = Reflect.get(target, key, reciver)
+      const res = Reflect.get(target, key, reciver);
       if (res === undefined) {
-        const p: any = createErrorProxy()
-        ;(target as any)[key] = p
-        return p
+        const p: any = createErrorProxy();
+        (target as any)[key] = p;
+        return p;
       }
 
-      return res
+      return res;
     },
-  })
+  });
 }
 
 export function mergeObjects(obj1: any, obj2: any, concatArrays = false) {
   // Recursively merge deeply nested objects.
-  const acc = Object.assign({}, obj1) // Prevent mutation of source object.
+  const acc = Object.assign({}, obj1); // Prevent mutation of source object.
   return Object.keys(obj2).reduce((acc, key) => {
     const left = obj1 ? obj1[key] : {},
-      right = obj2[key]
+      right = obj2[key];
     if (obj1 && obj1.hasOwnProperty(key) && isObject(right)) {
-      acc[key] = mergeObjects(left, right, concatArrays)
+      acc[key] = mergeObjects(left, right, concatArrays);
     } else if (concatArrays && Array.isArray(left) && Array.isArray(right)) {
-      acc[key] = left.concat(right)
+      acc[key] = left.concat(right);
     } else {
-      acc[key] = right
+      acc[key] = right;
     }
-    return acc
-  }, acc)
+    return acc;
+  }, acc);
 }
