@@ -1,4 +1,4 @@
 module.exports = {
-  presets: ['@vue/cli-plugin-babel/preset'],
-  plugins: [['@vue/babel-plugin-jsx', { mergeProps: false }]],
-}
+  presets: ["@vue/cli-plugin-babel/preset"],
+  plugins: [["@vue/babel-plugin-jsx", { mergeProps: false }]]
+};
