@@ -1,15 +1,15 @@
 module.exports = {
-  preset: '@vue/cli-plugin-unit-jest/presets/typescript-and-babel',
+  preset: "@vue/cli-plugin-unit-jest/presets/typescript-and-babel",
   moduleFileExtensions: [
-    'js',
-    'jsx',
-    'json',
-    'ts',
-    'tsx',
+    "js",
+    "jsx",
+    "json",
+    "ts",
+    "tsx",
     // tell Jest to handle *.vue files
-    'vue',
+    "vue"
   ],
   transform: {
-    '^.+\\.vue$': 'vue-jest',
-  },
-}
+    "^.+\\.vue$": "vue-jest"
+  }
+};
