@@ -5,79 +5,80 @@ import {
   Ref,
   watchEffect,
   watch,
-  shallowRef,
-} from 'vue'
+  shallowRef
+} from "vue";
 
-import Ajv, { Options } from 'ajv'
+import Ajv, { Options } from "ajv";
 
-import { Schema, SchemaTypes, Theme } from './types'
+import { Schema, SchemaTypes, Theme } from "./types";
 
-import SchemaItem from './SchemaItem'
-import { SchemaFormContextKey } from './context'
-import { validateFormData, ErrorSchema } from './validator'
+import SchemaItem from "./SchemaItem";
+import { SchemaFormContextKey } from "./context";
+import { validateFormData, ErrorSchema } from "./validator";
 
-type A = typeof SchemaItem
+type A = typeof SchemaItem;
 
 interface ContextRef {
   doValidate: () => {
-    errors: any[]
-    valid: boolean
-  }
+    errors: any[];
+    valid: boolean;
+  };
 }
 
 const defaultAjvOptions: Options = {
   allErrors: true,
-  jsonPointers: true,
-}
+  jsonPointers: true
+};
 
 export default defineComponent({
   props: {
     schema: {
       type: Object as PropType<Schema>,
-      required: true,
+      required: true
     },
     value: {
-      required: true,
+      required: true
     },
     onChange: {
       type: Function as PropType<(v: any) => void>,
-      required: true,
+      required: true
     },
     contextRef: {
-      type: Object as PropType<Ref<ContextRef | undefined>>,
+      type: Object as PropType<Ref<ContextRef | undefined>>
     },
     ajvOptions: {
-      type: Object as PropType<Options>,
+      type: Object as PropType<Options>
     },
     locale: {
       type: String,
-      default: 'zh',
+      default: "zh"
     },
     customValidate: {
-      type: Function as PropType<(data: any, errors: any) => void>,
-    },
+      type: Function as PropType<(data: any, errors: any) => void>
+    }
   },
-  name: 'SchemaForm',
+  name: "SchemaForm",
   setup(props, { slots, emit, attrs }) {
     const handleChange = (v: any) => {
-      props.onChange(v)
-    }
+      props.onChange(v);
+    };
 
     const context: any = {
-      SchemaItem,
+      SchemaItem
       // theme: props.theme,
-    }
+    };
 
-    const errorSchemaRef: Ref<ErrorSchema> = shallowRef({})
+    const errorSchemaRef: Ref<ErrorSchema> = shallowRef({});
 
-    const validatorRef: Ref<Ajv.Ajv> = shallowRef() as any
+    const validatorRef: Ref<Ajv.Ajv> = shallowRef() as any;
+    // const ajvOptions = props.ajvOptions;
 
     watchEffect(() => {
       validatorRef.value = new Ajv({
         ...defaultAjvOptions,
-        ...props.ajvOptions,
-      })
-    })
+        ...Object.assign({}, props.ajvOptions)
+      });
+    });
 
     watch(
       () => props.contextRef,
@@ -85,7 +86,7 @@ export default defineComponent({
         if (props.contextRef) {
           props.contextRef.value = {
             doValidate() {
-              console.log('--------->')
+              console.log("--------->");
 
               // const valid = validatorRef.value.validate(
               //   props.schema,
@@ -97,25 +98,25 @@ export default defineComponent({
                 props.value,
                 props.schema,
                 props.locale,
-                props.customValidate,
-              )
+                props.customValidate
+              );
 
-              errorSchemaRef.value = result.errorSchema
+              errorSchemaRef.value = result.errorSchema;
 
-              return result
-            },
-          }
+              return result;
+            }
+          };
         }
       },
       {
-        immediate: true,
-      },
-    )
+        immediate: true
+      }
+    );
 
-    provide(SchemaFormContextKey, context)
+    provide(SchemaFormContextKey, context);
 
     return () => {
-      const { schema, value } = props
+      const { schema, value } = props;
       return (
         <SchemaItem
           schema={schema}
@@ -124,7 +125,7 @@ export default defineComponent({
           onChange={handleChange}
           errorSchema={errorSchemaRef.value || {}}
         />
-      )
-    }
-  },
-})
+      );
+    };
+  }
+});
