@@ -1,58 +1,58 @@
-import { computed, defineComponent, PropType } from 'vue'
+import { computed, defineComponent, PropType } from "vue";
 
-import { Schema, SchemaTypes, FiledPropsDefine } from './types'
+import { Schema, SchemaTypes, FiledPropsDefine } from "./types";
 // import StringField from './fields/StringField'
-import StringField from './fields/StringField'
-import NumberField from './fields/NumberField'
+import StringField from "./fields/StringField";
+import NumberField from "./fields/NumberField";
 
-import ObjectField from './fields/ObjectField'
-import ArrayField from './fields/ArrayField'
+import ObjectField from "./fields/ObjectField";
+import ArrayField from "./fields/ArrayField";
 
-import { retrieveSchema } from './utils'
+import { retrieveSchema } from "./utils";
 
 export default defineComponent({
-  name: 'SchemaItem',
+  name: "SchemaItem",
   props: FiledPropsDefine,
   setup(props) {
     const retrievedSchemaRef = computed(() => {
-      const { schema, rootSchema, value } = props
-      return retrieveSchema(schema, rootSchema, value)
-    })
+      const { schema, rootSchema, value } = props;
+      return retrieveSchema(schema, rootSchema, value);
+    });
 
     return () => {
-      const { schema, rootSchema, value } = props
+      const { schema, rootSchema, value } = props;
 
-      const retrievedSchema = retrievedSchemaRef.value
+      const retrievedSchema = retrievedSchemaRef.value;
 
       // TODO: 如果type没有指定，我们需要猜测这个type
 
-      const type = schema.type
+      const type = schema["type"];
 
-      let Component: any
+      let Component: any;
 
       switch (type) {
         case SchemaTypes.STRING: {
-          Component = StringField
-          break
+          Component = StringField;
+          break;
         }
         case SchemaTypes.NUMBER: {
-          Component = NumberField
-          break
+          Component = NumberField;
+          break;
         }
         case SchemaTypes.OBJECT: {
-          Component = ObjectField
-          break
+          Component = ObjectField;
+          break;
         }
         case SchemaTypes.ARRAY: {
-          Component = ArrayField
-          break
+          Component = ArrayField;
+          break;
         }
         default: {
-          console.warn(`${type} is not supported`)
+          console.warn(`${type} is not supported`);
         }
       }
 
-      return <Component {...props} schema={retrievedSchema} />
-    }
+      return <Component {...props} schema={retrievedSchema} />;
+    };
   },
-})
+});
