@@ -1,68 +1,68 @@
-import { defineComponent, PropType } from 'vue'
-import { createUseStyles } from 'vue-jss'
+import { defineComponent, PropType } from "vue";
+import { createUseStyles } from "vue-jss";
 
-import { FiledPropsDefine, Schema, SelectionWidgetNames } from '../types'
+import { FiledPropsDefine, Schema, SelectionWidgetNames } from "../types";
 
-import { useVJSFContext } from '../context'
-import { getWidget } from '../theme'
+import { useVJSFContext } from "../context";
+import { getWidget } from "../theme";
 
 // import SelectionWidget from '../widgets/Selection'
 
 const useStyles = createUseStyles({
   container: {
-    border: '1px solid #eee',
+    border: "1px solid #eee"
   },
   actions: {
-    background: '#eee',
+    background: "#eee",
     padding: 10,
-    textAlign: 'right',
+    textAlign: "right"
   },
   action: {
-    '& + &': {
-      marginLeft: 10,
-    },
+    "& + &": {
+      marginLeft: 10
+    }
   },
   content: {
-    padding: 10,
-  },
-})
+    padding: 10
+  }
+});
 
 const ArrayItemWrapper = defineComponent({
-  name: 'ArrayItemWrapper',
+  name: "ArrayItemWrapper",
   props: {
     onAdd: {
       type: Function as PropType<(index: number) => void>,
-      required: true,
+      required: true
     },
     onDelete: {
       type: Function as PropType<(index: number) => void>,
-      required: true,
+      required: true
     },
     onUp: {
       type: Function as PropType<(index: number) => void>,
-      required: true,
+      required: true
     },
     onDown: {
       type: Function as PropType<(index: number) => void>,
-      required: true,
+      required: true
     },
     index: {
       type: Number,
-      required: true,
-    },
+      required: true
+    }
   },
   setup(props, { slots }) {
-    const classesRef = useStyles()
+    const classesRef = useStyles();
 
-    const context = useVJSFContext()
+    const context = useVJSFContext();
 
-    const handleAdd = () => props.onAdd(props.index)
-    const handleDown = () => props.onDown(props.index)
-    const handleUp = () => props.onUp(props.index)
-    const handleDelete = () => props.onDelete(props.index)
+    const handleAdd = () => props.onAdd(props.index);
+    const handleDown = () => props.onDown(props.index);
+    const handleUp = () => props.onUp(props.index);
+    const handleDelete = () => props.onDelete(props.index);
 
     return () => {
-      const classes = classesRef.value
+      const classes = classesRef.value;
 
       return (
         <div class={classes.container}>
@@ -82,10 +82,10 @@ const ArrayItemWrapper = defineComponent({
           </div>
           <div class={classes.content}>{slots.default && slots.default()}</div>
         </div>
-      )
-    }
-  },
-})
+      );
+    };
+  }
+});
 
 /**
  * {
@@ -104,76 +104,76 @@ const ArrayItemWrapper = defineComponent({
  * }
  */
 export default defineComponent({
-  name: 'ArrayField',
+  name: "ArrayField",
   props: FiledPropsDefine,
   setup(props) {
-    const context = useVJSFContext()
+    const context = useVJSFContext();
 
     const handleArrayItemChange = (v: any, index: number) => {
-      const { value } = props
-      const arr = Array.isArray(value) ? value : []
+      const { value } = props;
+      const arr = Array.isArray(value) ? value : [];
 
-      arr[index] = v
+      arr[index] = v;
 
-      props.onChange(arr)
-    }
+      props.onChange(arr);
+    };
 
     const handleAdd = (index: number) => {
-      const { value } = props
-      const arr = Array.isArray(value) ? value : []
+      const { value } = props;
+      const arr = Array.isArray(value) ? value : [];
 
-      arr.splice(index + 1, 0, undefined)
+      arr.splice(index + 1, 0, undefined);
 
-      props.onChange(arr)
-    }
+      props.onChange(arr);
+    };
 
     const handleDelete = (index: number) => {
-      const { value } = props
-      const arr = Array.isArray(value) ? value : []
+      const { value } = props;
+      const arr = Array.isArray(value) ? value : [];
 
-      arr.splice(index, 1)
+      arr.splice(index, 1);
 
-      props.onChange(arr)
-    }
+      props.onChange(arr);
+    };
 
     const handleUp = (index: number) => {
-      if (index === 0) return
-      const { value } = props
-      const arr = Array.isArray(value) ? value : []
+      if (index === 0) return;
+      const { value } = props;
+      const arr = Array.isArray(value) ? value : [];
 
-      const item = arr.splice(index, 1)
-      arr.splice(index - 1, 0, item[0])
+      const item = arr.splice(index, 1);
+      arr.splice(index - 1, 0, item[0]);
 
-      props.onChange(arr)
-    }
+      props.onChange(arr);
+    };
 
     const handleDown = (index: number) => {
-      const { value } = props
-      const arr = Array.isArray(value) ? value : []
+      const { value } = props;
+      const arr = Array.isArray(value) ? value : [];
 
-      if (index === arr.length - 1) return
+      if (index === arr.length - 1) return;
 
-      const item = arr.splice(index, 1)
-      arr.splice(index + 1, 0, item[0])
+      const item = arr.splice(index, 1);
+      arr.splice(index + 1, 0, item[0]);
 
-      props.onChange(arr)
-    }
+      props.onChange(arr);
+    };
 
-    const SelectionWidgetRef = getWidget(SelectionWidgetNames.SelectionWidget)
+    const SelectionWidgetRef = getWidget(SelectionWidgetNames.SelectionWidget);
 
     return () => {
       // const SelectionWidget = context.theme.widgets.SelectionWidget
-      const SelectionWidget = SelectionWidgetRef.value
-      const { schema, rootSchema, value, errorSchema } = props
+      const SelectionWidget = SelectionWidgetRef.value;
+      const { schema, rootSchema, value, errorSchema } = props;
 
-      const SchemaItem = context.SchemaItem
+      const SchemaItem = context.SchemaItem;
 
-      const isMultiType = Array.isArray(schema.items)
-      const isSelect = schema.items && (schema.items as any).enum
+      const isMultiType = Array.isArray(schema["items"]);
+      const isSelect = schema["items"] && (schema["items"] as any).enum;
 
       if (isMultiType) {
-        const items: Schema[] = schema.items as any
-        const arr = Array.isArray(value) ? value : []
+        const items: Schema[] = schema["items"] as any;
+        const arr = Array.isArray(value) ? value : [];
         return items.map((s: Schema, index: number) => (
           <SchemaItem
             schema={s}
@@ -183,9 +183,9 @@ export default defineComponent({
             errorSchema={errorSchema[index] || {}}
             onChange={(v: any) => handleArrayItemChange(v, index)}
           />
-        ))
+        ));
       } else if (!isSelect) {
-        const arr = Array.isArray(value) ? value : []
+        const arr = Array.isArray(value) ? value : [];
 
         return arr.map((v: any, index: number) => {
           return (
@@ -197,7 +197,7 @@ export default defineComponent({
               onUp={handleUp}
             >
               <SchemaItem
-                schema={schema.items as Schema}
+                schema={schema["items"] as Schema}
                 errorSchema={errorSchema[index] || {}}
                 value={v}
                 key={index}
@@ -205,14 +205,14 @@ export default defineComponent({
                 onChange={(v: any) => handleArrayItemChange(v, index)}
               />
             </ArrayItemWrapper>
-          )
-        })
+          );
+        });
       } else {
-        const enumOptions = (schema as any).items.enum
+        const enumOptions = (schema as any).items.enum;
         const options = enumOptions.map((e: any) => ({
           key: e,
-          value: e,
-        }))
+          value: e
+        }));
         return (
           <SelectionWidget
             onChange={props.onChange}
@@ -221,8 +221,8 @@ export default defineComponent({
             errors={errorSchema.__errors}
             schema={schema}
           />
-        )
+        );
       }
-    }
-  },
-})
+    };
+  }
+});
