@@ -1,53 +1,53 @@
-import { defineComponent, DefineComponent } from 'vue'
+import { defineComponent, DefineComponent } from "vue";
 
-import { FiledPropsDefine, CommonFieldType } from '../types'
-import { isObject } from '../utils'
-import { SchemaFormContextKey, useVJSFContext } from '../context'
+import { FiledPropsDefine, CommonFieldType } from "../types";
+import { isObject } from "../utils";
+import { SchemaFormContextKey, useVJSFContext } from "../context";
 
 // import SchemaItem from '../SchemaItem'
 
 // console.log(SchemaItem)
 
 const schema = {
-  type: 'object',
+  type: "object",
   properties: {
     name: {
-      type: 'string',
+      type: "string"
     },
     age: {
-      type: 'number',
-    },
-  },
-}
+      type: "number"
+    }
+  }
+};
 
-type A = DefineComponent<typeof FiledPropsDefine, {}, {}>
+type A = DefineComponent<typeof FiledPropsDefine, {}, {}>;
 
 export default defineComponent({
-  name: 'ObjectField',
+  name: "ObjectField",
   props: FiledPropsDefine,
   setup(props) {
-    const context = useVJSFContext()
+    const context = useVJSFContext();
 
     const handleObjectFieldChange = (key: string, v: any) => {
-      const value: any = isObject(props.value) ? props.value : {}
+      const value: any = isObject(props.value) ? props.value : {};
 
       if (v === undefined) {
-        delete value[key]
+        delete value[key];
       } else {
-        value[key] = v
+        value[key] = v;
       }
 
-      props.onChange(value)
-    }
+      props.onChange(value);
+    };
 
     return () => {
-      const { schema, rootSchema, value, errorSchema } = props
+      const { schema, rootSchema, value, errorSchema } = props;
 
-      const { SchemaItem } = context
+      const { SchemaItem } = context;
 
-      const properties = schema.properties || {}
+      const properties = schema["properties"] || {};
 
-      const currentValue: any = isObject(value) ? value : {}
+      const currentValue: any = isObject(value) ? value : {};
 
       return Object.keys(properties).map((k: string, index: number) => (
         <SchemaItem
@@ -58,7 +58,7 @@ export default defineComponent({
           key={index}
           onChange={(v: any) => handleObjectFieldChange(k, v)}
         />
-      ))
-    }
-  },
-})
+      ));
+    };
+  }
+});
