@@ -1,23 +1,23 @@
-import { defineComponent } from 'vue'
+import { defineComponent } from "vue";
 
-import { FiledPropsDefine, CommonWidgetNames } from '../types'
-import { getWidget } from '../theme'
+import { FiledPropsDefine, CommonWidgetNames } from "../types";
+import { getWidget } from "../theme";
 
 export default defineComponent({
-  name: 'StringFeild',
+  name: "StringFeild",
   props: FiledPropsDefine,
   setup(props) {
     const handleChange = (v: string) => {
       // console.log(e)
-      props.onChange(v)
-    }
+      props.onChange(v);
+    };
 
-    const TextWidgetRef = getWidget(CommonWidgetNames.TextWidget)
+    const TextWidgetRef = getWidget(CommonWidgetNames.TextWidget);
 
     return () => {
-      const { rootSchema, errorSchema, ...rest } = props
+      const { rootSchema, errorSchema, ...rest } = props;
 
-      const TextWidget = TextWidgetRef.value
+      const TextWidget = TextWidgetRef.value;
 
       return (
         <TextWidget
@@ -25,11 +25,11 @@ export default defineComponent({
           errors={errorSchema.__errors}
           onChange={handleChange}
         />
-      )
+      );
 
       // return (
       //   <input type="text" value={props.value as any} onInput={handleChange} />
       // )
-    }
-  },
-})
+    };
+  }
+});
