@@ -1,47 +1,47 @@
-import { defineComponent } from 'vue'
-import { CommonWidgetPropsDefine, CommonWidgetDefine } from '../types'
+import { defineComponent } from "vue";
+import { CommonWidgetPropsDefine, CommonWidgetDefine } from "../types";
 
-import { createUseStyles } from 'vue-jss'
+import { createUseStyles } from "vue-jss";
 
 const useStyles = createUseStyles({
   container: {},
   label: {
-    display: 'block',
-    color: '#777',
+    display: "block",
+    color: "#777"
   },
   errorText: {
-    color: 'red',
+    color: "red",
     fontSize: 12,
-    margin: '5px 0',
+    margin: "5px 0",
     padding: 0,
-    paddingLeft: 20,
-  },
-})
+    paddingLeft: 20
+  }
+});
 
 const FormItem = defineComponent({
-  name: 'FormItem',
+  name: "FormItem",
   props: CommonWidgetPropsDefine,
   setup(props, { slots }) {
-    const classesRef = useStyles()
+    const classesRef = useStyles();
     return () => {
-      const { schema, errors } = props
-      const classes = classesRef.value
+      const { schema, errors } = props;
+      const classes = classesRef.value;
       return (
         <div class={classes.container}>
-          <label class={classes.label}>{schema.title}</label>
+          <label class={classes.label}>{schema["title"]}</label>
           {slots.default && slots.default()}
           <ul class={classes.errorText}>
-            {errors?.map((err) => (
+            {errors?.map(err => (
               <li>{err}</li>
             ))}
           </ul>
         </div>
-      )
-    }
-  },
-})
+      );
+    };
+  }
+});
 
-export default FormItem
+export default FormItem;
 
 // HOC: Higher Order Component: 高阶组件
 export function withFormItem(Widget: any) {
@@ -54,8 +54,8 @@ export function withFormItem(Widget: any) {
           <FormItem {...props}>
             <Widget {...props} {...attrs} />
           </FormItem>
-        )
-      }
-    },
-  }) as any
+        );
+      };
+    }
+  }) as any;
 }
