@@ -1,18 +1,18 @@
-import { CommonWidgetPropsDefine, CommonWidgetDefine } from '../types'
-import { defineComponent } from 'vue'
+import { CommonWidgetPropsDefine, CommonWidgetDefine } from "../types";
+import { defineComponent } from "vue";
 
-import { withFormItem } from './FormItem'
+import { withFormItem } from "./FormItem";
 
 const TextWidget: CommonWidgetDefine = withFormItem(
   defineComponent({
-    name: 'TextWidget',
+    name: "TextWidget",
     props: CommonWidgetPropsDefine,
     setup(props) {
       const handleChange = (e: any) => {
-        const value = e.target.value
-        e.target.value = props.value
-        props.onChange(value)
-      }
+        const value = e.target.value;
+        e.target.value = props.value;
+        props.onChange(value);
+      };
       return () => {
         return (
           <input
@@ -20,10 +20,10 @@ const TextWidget: CommonWidgetDefine = withFormItem(
             value={props.value as any}
             onInput={handleChange}
           />
-        )
-      }
-    },
-  }),
-)
+        );
+      };
+    }
+  })
+);
 
-export default TextWidget
+export default TextWidget;
