@@ -4,49 +4,49 @@ import {
   inject,
   PropType,
   provide,
-  ComputedRef,
-} from 'vue'
+  ComputedRef
+} from "vue";
 import {
   Theme,
   SelectionWidgetNames,
   CommonWidgetNames,
-  SelectionWidgetDefine,
-} from './types'
+  SelectionWidgetDefine
+} from "./types";
 
-const THEME_PROVIDER_KEY = Symbol()
+const THEME_PROVIDER_KEY = Symbol();
 
 const ThemeProvider = defineComponent({
-  name: 'VJSFThemeProvider',
+  name: "VJSFThemeProvider",
   props: {
     theme: {
       type: Object as PropType<Theme>,
-      required: true,
-    },
+      required: true
+    }
   },
   setup(props, { slots }) {
-    const context = computed(() => props.theme)
+    const context = computed(() => props.theme);
 
-    provide(THEME_PROVIDER_KEY, context)
+    provide(THEME_PROVIDER_KEY, context);
 
-    return () => slots.default && slots.default()
-  },
-})
+    return () => slots.default && slots.default();
+  }
+});
 
 export function getWidget<T extends SelectionWidgetNames | CommonWidgetNames>(
-  name: T,
+  name: T
 ) {
   const context: ComputedRef<Theme> | undefined = inject<ComputedRef<Theme>>(
-    THEME_PROVIDER_KEY,
-  )
+    THEME_PROVIDER_KEY
+  );
   if (!context) {
-    throw new Error('vjsf theme required')
+    throw new Error("vjsf theme required");
   }
 
   const widgetRef = computed(() => {
-    return context.value.widgets[name]
-  })
+    return context.value.widgets[name];
+  });
 
-  return widgetRef
+  return widgetRef;
 }
 
-export default ThemeProvider
+export default ThemeProvider;
