@@ -1,53 +1,53 @@
 // Node.js require:
-const Ajv = require('ajv')
-const localize = require('ajv-i18n')
+const Ajv = require("ajv");
+const localize = require("ajv-i18n");
 
 const schema = {
-  type: 'object',
+  type: "object",
   properties: {
     name: {
-      type: 'string',
+      type: "string",
       // test: false,
       errorMessage: {
-        type: '必须是字符串',
-        minLength: '长度不能小于10',
+        type: "必须是字符串",
+        minLength: "长度不能小于10",
       },
       // format: 'test',
       minLength: 10,
     },
     age: {
-      type: 'number',
+      type: "number",
     },
     pets: {
-      type: 'array',
+      type: "array",
       items: [
         {
-          type: 'string',
+          type: "string",
           maxLength: 2,
         },
         {
-          type: 'number',
+          type: "number",
         },
       ],
     },
     isWorker: {
-      type: 'boolean',
+      type: "boolean",
     },
   },
-  required: ['name', 'age'],
-}
+  required: ["name", "age"],
+};
 
-const ajv = new Ajv({ allErrors: true, jsonPointers: true }) // options can be passed, e.g. {allErrors: true}
+const ajv = new Ajv({ allErrors: true, jsonPointers: true }); // options can be passed, e.g. {allErrors: true}
 // ajv.addFormat('test', (data) => {
 //   console.log(data, '------------')
 //   return data === 'haha'
 // })
-require('ajv-errors')(ajv)
-ajv.addKeyword('test', {
+require("ajv-errors")(ajv);
+ajv.addKeyword("test", {
   macro() {
     return {
       minLength: 10,
-    }
+    };
   },
   // compile(sch, parentSchema) {
   //   console.log(sch, parentSchema)
@@ -72,15 +72,15 @@ ajv.addKeyword('test', {
 
   //   return false
   // },
-})
-const validate = ajv.compile(schema)
+});
+const validate = ajv.compile(schema);
 const valid = validate({
-  name: '12',
+  name: "12",
   age: 18,
-  pets: ['mimi', 12],
+  pets: ["mimi", 12],
   isWorker: true,
-})
+});
 if (!valid) {
-  localize.zh(validate.errors)
-  console.log(validate.errors)
+  localize.zh(validate.errors);
+  console.log(validate.errors);
 }
