@@ -1,143 +1,143 @@
-import { defineComponent, ref, Ref, reactive, watchEffect } from 'vue'
-import { createUseStyles } from 'vue-jss'
+import { defineComponent, ref, Ref, reactive, watchEffect } from "vue";
+import { createUseStyles } from "vue-jss";
 
-import MonacoEditor from './components/MonacoEditor'
+import MonacoEditor from "./components/MonacoEditor";
 
-import demos from './demos'
+import demos from "./demos";
 
-import SchemaForm, { ThemeProvider } from '../lib'
-import themeDefault from '../lib/theme-default'
+import SchemaForm, { ThemeProvider } from "../lib";
+import themeDefault from "../lib/theme-default";
 
 // TODO: 在lib中export
-type Schema = any
-type UISchema = any
+type Schema = any;
+type UISchema = any;
 
 function toJson(data: any) {
-  return JSON.stringify(data, null, 2)
+  return JSON.stringify(data, null, 2);
 }
 
 const useStyles = createUseStyles({
   container: {
-    display: 'flex',
-    flexDirection: 'column',
-    height: '100%',
-    width: '1200px',
-    margin: '0 auto',
+    display: "flex",
+    flexDirection: "column",
+    height: "100%",
+    width: "1200px",
+    margin: "0 auto"
   },
   menu: {
-    marginBottom: 20,
+    marginBottom: 20
   },
   code: {
     width: 700,
-    flexShrink: 0,
+    flexShrink: 0
   },
   codePanel: {
     minHeight: 400,
-    marginBottom: 20,
+    marginBottom: 20
   },
   uiAndValue: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    '& > *': {
-      width: '46%',
-    },
+    display: "flex",
+    justifyContent: "space-between",
+    "& > *": {
+      width: "46%"
+    }
   },
   content: {
-    display: 'flex',
+    display: "flex"
   },
   form: {
-    padding: '0 20px',
-    flexGrow: 1,
+    padding: "0 20px",
+    flexGrow: 1
   },
   menuButton: {
-    appearance: 'none',
+    appearance: "none",
     borderWidth: 0,
-    backgroundColor: 'transparent',
-    cursor: 'pointer',
-    display: 'inline-block',
+    backgroundColor: "transparent",
+    cursor: "pointer",
+    display: "inline-block",
     padding: 15,
     borderRadius: 5,
-    '&:hover': {
-      background: '#efefef',
-    },
+    "&:hover": {
+      background: "#efefef"
+    }
   },
   menuSelected: {
-    background: '#337ab7',
-    color: '#fff',
-    '&:hover': {
-      background: '#337ab7',
-    },
-  },
-})
+    background: "#337ab7",
+    color: "#fff",
+    "&:hover": {
+      background: "#337ab7"
+    }
+  }
+});
 
 export default defineComponent({
   setup() {
-    const selectedRef: Ref<number> = ref(0)
+    const selectedRef: Ref<number> = ref(0);
 
     const demo: {
-      schema: Schema | null
-      data: any
-      uiSchema: UISchema | null
-      schemaCode: string
-      dataCode: string
-      uiSchemaCode: string
-      customValidate: ((d: any, e: any) => void) | undefined
+      schema: Schema | null;
+      data: any;
+      uiSchema: UISchema | null;
+      schemaCode: string;
+      dataCode: string;
+      uiSchemaCode: string;
+      customValidate: ((d: any, e: any) => void) | undefined;
     } = reactive({
       schema: null,
       data: {},
       uiSchema: {},
-      schemaCode: '',
-      dataCode: '',
-      uiSchemaCode: '',
-      customValidate: undefined,
-    })
+      schemaCode: "",
+      dataCode: "",
+      uiSchemaCode: "",
+      customValidate: undefined
+    });
 
     watchEffect(() => {
-      const index = selectedRef.value
-      const d: any = demos[index]
-      demo.schema = d.schema
-      demo.data = d.default
-      demo.uiSchema = d.uiSchema
-      demo.schemaCode = toJson(d.schema)
-      demo.dataCode = toJson(d.default)
-      demo.uiSchemaCode = toJson(d.uiSchema)
-      demo.customValidate = d.customValidate
-    })
+      const index = selectedRef.value;
+      const d: any = demos[index];
+      demo.schema = d.schema;
+      demo.data = d.default;
+      demo.uiSchema = d.uiSchema;
+      demo.schemaCode = toJson(d.schema);
+      demo.dataCode = toJson(d.default);
+      demo.uiSchemaCode = toJson(d.uiSchema);
+      demo.customValidate = d.customValidate;
+    });
 
-    const methodRef: Ref<any> = ref()
+    const methodRef: Ref<any> = ref();
 
-    const classesRef = useStyles()
+    const classesRef = useStyles();
 
     const handleChange = (v: any) => {
-      demo.data = v
-      demo.dataCode = toJson(v)
-    }
+      demo.data = v;
+      demo.dataCode = toJson(v);
+    };
 
     function handleCodeChange(
-      filed: 'schema' | 'data' | 'uiSchema',
-      value: string,
+      filed: "schema" | "data" | "uiSchema",
+      value: string
     ) {
       try {
-        const json = JSON.parse(value)
-        demo[filed] = json
-        ;(demo as any)[`${filed}Code`] = value
+        const json = JSON.parse(value);
+        demo[filed] = json;
+        (demo as any)[`${filed}Code`] = value;
       } catch (err) {
         // some thing
       }
     }
 
-    const handleSchemaChange = (v: string) => handleCodeChange('schema', v)
-    const handleDataChange = (v: string) => handleCodeChange('data', v)
-    const handleUISchemaChange = (v: string) => handleCodeChange('uiSchema', v)
+    const handleSchemaChange = (v: string) => handleCodeChange("schema", v);
+    const handleDataChange = (v: string) => handleCodeChange("data", v);
+    const handleUISchemaChange = (v: string) => handleCodeChange("uiSchema", v);
 
-    const contextRef = ref()
-    const nameRef = ref()
+    const contextRef = ref();
+    const nameRef = ref();
 
     return () => {
-      const classes = classesRef.value
-      const selected = selectedRef.value
+      const classes = classesRef.value;
+      const selected = selectedRef.value;
 
-      console.log(methodRef, nameRef)
+      console.log(methodRef, nameRef);
 
       return (
         // <StyleThemeProvider>
@@ -150,7 +150,7 @@ export default defineComponent({
                 <button
                   class={{
                     [classes.menuButton]: true,
-                    [classes.menuSelected]: index === selected,
+                    [classes.menuSelected]: index === selected
                   }}
                   onClick={() => (selectedRef.value = index)}
                 >
@@ -210,7 +210,7 @@ export default defineComponent({
         </div>
         // </VJSFThemeProvider>
         // </StyleThemeProvider>
-      )
-    }
-  },
-})
+      );
+    };
+  }
+});
