@@ -1,25 +1,25 @@
 export default {
-  name: 'Demo',
+  name: "Demo",
   schema: {
-    type: 'object',
+    type: "object",
     properties: {
       pass1: {
-        type: 'string',
+        type: "string",
         minLength: 10,
-        title: 'password',
+        title: "password"
       },
       pass2: {
-        type: 'string',
+        type: "string",
         minLength: 10,
-        title: 're try password',
-      },
-    },
+        title: "re try password"
+      }
+    }
   },
   customValidate(data: any, errors: any) {
     if (data.pass1 !== data.pass2) {
-      errors.pass2.addError('密码必须相同')
+      errors.pass2.addError("密码必须相同");
     }
   },
   uiSchema: {},
-  default: 1,
-}
+  default: 1
+};
