@@ -1,29 +1,29 @@
-import { mount, shallowMount } from '@vue/test-utils'
-import { defineComponent, h } from 'vue'
+import { mount, shallowMount } from "@vue/test-utils";
+import { defineComponent, h } from "vue";
 
-import JsonSchemaForm, { NumberFiled } from '../../lib'
+import JsonSchemaForm, { NumberFiled } from "../../lib";
 
-describe('JsonSchemaFrom', () => {
-  it('should render correct number field', async () => {
-    let value = ''
+describe("JsonSchemaFrom", () => {
+  it("should render correct number field", async () => {
+    let value = "";
     const wrapper = mount(JsonSchemaForm, {
       props: {
         schema: {
-          type: 'number',
+          type: "number",
         },
         value: value,
         onChange: (v) => {
-          value = v
+          value = v;
         },
       },
-    })
+    });
 
-    const numberFiled = wrapper.findComponent(NumberFiled)
-    expect(numberFiled.exists()).toBeTruthy()
+    const numberFiled = wrapper.findComponent(NumberFiled);
+    expect(numberFiled.exists()).toBeTruthy();
     // await numberFiled.props('onChange')('123')
-    const input = numberFiled.find('input')
-    input.element.value = '123'
-    input.trigger('input')
-    expect(value).toBe(123)
-  })
-})
+    const input = numberFiled.find("input");
+    input.element.value = "123";
+    input.trigger("input");
+    expect(value).toBe(123);
+  });
+});
